+++ app/styles/borderStyles.tsx
@@ -1,3 +1,9 @@
 import { cn } from '@/lib/utils';
 
-export const cnBorder = cn('border rounded-xl shadow');
+export const cnBorderNone = 'border-none';
+
+export const cnBorder = cn('border shadow');
+
+export const cnBorder2 = 'border-2 shadow';
+
+export const cnBorderBottom4 = 'border-b-4 shadow';
