+++ app/styles/imageStyles.tsx
@@ -1,0 +1,7 @@
+import { cn } from '@/lib/utils';
+
+export const cnBigImage = cn(
+  'max-h-[92px] min-h-[92px] min-w-[92px] max-w-[92px]',
+  'sm:max-h-[132px] sm:min-h-[132px] sm:min-w-[132px] sm:max-w-[132px]',
+  'md:max-h-[124px] md:min-h-[124px] md:min-w-[124px] md:max-w-[124px]'
+);
