--- app/styles/lineThroughStyles.tsx
@@ -1,3 +1,0 @@
-export const lineThroughItem = (date: string) => {
-  return date.includes('present') ? '' : 'line-through text-muted-foreground';
-};
