+++ app/styles/sizeStyles.tsx
@@ -1,7 +1,16 @@
 import { cn } from '@/lib/utils';
 
-export const sizeIcon = cn('w-auto', 'h-3', 'sm:h-4', 'mb:h-5');
+export const cnSizeIcon = cn('w-auto', 'h-3', 'sm:h-4', 'mb:h-5');
 
-export const sizeMiddleIcon = cn('w-auto  h-7', 'sm:h-8', 'mb:h-15', 'lg:h-18');
+export const cnSizeMiddleIcon = cn(
+  'w-auto  h-7',
+  'sm:h-8',
+  'mb:h-15',
+  'lg:h-18'
+);
 
-export const sizeBigIcon = cn('h-10 w-auto', 'xs:h-12', 'sm:h-14', 'mb:h-18');
+export const cnSizeBigIcon = cn('h-10 w-auto', 'xs:h-12', 'sm:h-14', 'mb:h-18');
+
+export const cnSizeAuto = 'w-auto h-auto';
+
+export const cnSizeFull = 'w-full h-full';
