+++ app/styles/translateStyles.tsx
@@ -1,0 +1,1 @@
+export const cnLittleTranslateSm = 'sm:-translate-y-0.5';
